+++ src/submission-service.ts
@@ -36,6 +36,7 @@ import {
 import { deleteAutoSaveData } from './auto-save-service'
 import { downloadFormSubmission } from './services/api/submissions'
 import { getSubmissionAttachmentDetails } from './attachments-service'
+import getPostSubmissionAttachmentURLs from './services/api/post-submission-attachment-urls'
 
 let _isProcessingPendingQueue = false
 
@@ -602,4 +603,5 @@ export {
   PendingQueueListener,
   PendingQueueAction,
   getSubmissionData,
+  getPostSubmissionAttachmentURLs,
 }